feat(advanced-techniques): add new API examples and update enum implementationThis commit adds new examples in the `API` package, including `ArrayListAPI` and `DateAPI` to demonstrate the usage of ArrayList and Date functionalities in Java. It also updates the `Enum_Direction` implementation to include string names for each enum constant and enhances `Enum_Test` to utilize the new `getName` method.
Additionally, the commit moves `Enum_Test` and `Enum_Direction` classes to a new
package `part1_enum` to better organize the project structure. Furthermore, it
introduces a new class `InitDemo` and its associated annotation `MyAnnotation` in
the `part2_annotation` package to illustrate the usage of annotations at runtime.

Other improvements include the addition of various string manipulation examples
in `StringAPI` and a demonstration of reflection capabilities in the `Test` class
within the `part2_annotation` package.